Well, shit broke with that commit. As expected, honestly. Fuck Visual Studio and it's merge tool.
It works now..I hope anyway.

+++ Raunstrup/Raunstrup.Core/Controllers/DraftController.cs
@@ -70,7 +70,7 @@ namespace Raunstrup.Core.Controllers
             _currentDraft.Description = description;
         }
 
-        public void SetDiscountPercentage(decimal percentage)
+        public void SetDiscountPercentage(double percentage)
         {
             _currentDraft.DiscountPercentage = percentage;
         }
